feat: 이미지 업로드 라우팅
- POST '/image'
